refactor: BookmarkProvideService 를 최적화
