+++ src/main/java/toy/subscribe/bookmark/service/BookmarkProvideService.java
@@ -1,9 +1,26 @@
 package toy.subscribe.bookmark.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import toy.subscribe.bookmark.dto.BookmarkDto;
+import toy.subscribe.bookmark.repository.BookmarkRepository;
+import toy.subscribe.configs.http.log.repository.HttpLogRepository;
 import toy.subscribe.configs.http.wrapper.HttpResponseWrapper;
 
-public interface BookmarkProvideService {
-    HttpResponseWrapper<BookmarkDto> provideBookmarkWrapper(Pageable pageable, String category, String title);
+@Service
+@RequiredArgsConstructor
+public class BookmarkProvideService {
+    private final BookmarkRepository bookmarkRepository;
+    private final HttpLogRepository requestLogRepository;
+
+    @Transactional(readOnly = true)
+    public HttpResponseWrapper<BookmarkDto> provideBookmarkWrapper(Pageable pageable, String category, String title) {
+        return HttpResponseWrapper.<BookmarkDto>builder()
+                                  .pages(bookmarkRepository.findPage(pageable, category, title))
+                                  .visitorsOfDay(requestLogRepository.findDau())
+                                  .visitorsOfReduce(requestLogRepository.findCumulativeVisitors())
+                                  .build();
+    }
 }
